# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/campaigns": {
      "post": {
        "description": "Create a new advertising campaign (Brand role required)",
        "operationId": "CampaignsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "61f8d3c97e1d2a001f9a5e8c",
                    "title": "Summer Gaming Promotion",
                    "description": "Promote our new gaming products during your streams",
                    "brandId": "60d21b4667d0d8992e610c85",
                    "budget": 1000,
                    "mediaUrl": "https://example.com/media/campaign123.mp4",
                    "mediaType": "video",
                    "status": "active",
                    "categories": [
                      "Gaming",
                      "Technology"
                    ],
                    "createdAt": "2025-06-25T12:00:00.000Z",
                    "updatedAt": "2025-06-25T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input data"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden - Must be a brand to create campaigns"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new campaign",
        "tags": [
          "campaigns"
        ]
      },
      "get": {
        "description": "Retrieve all campaigns with optional filtering",
        "operationId": "CampaignsController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter campaigns by category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter campaigns by status",
            "schema": {
              "enum": [
                "draft",
                "active",
                "paused",
                "completed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of campaigns",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": {
                      "id": "61f8d3c97e1d2a001f9a5e8c",
                      "title": "Summer Gaming Promotion",
                      "brandId": "60d21b4667d0d8992e610c85",
                      "status": "active",
                      "budget": 1000,
                      "mediaUrl": "https://example.com/media/campaign123.mp4",
                      "mediaType": "video"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get all campaigns",
        "tags": [
          "campaigns"
        ]
      }
    },
...
    "/api/v1/campaigns/{id}": {
      "get": {
        "operationId": "CampaignsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "adminAccess",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "campaigns"
        ]
      },
      "put": {
        "operationId": "CampaignsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "campaigns"
        ]
      },
      "delete": {
        "operationId": "CampaignsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/join": {
      "post": {
        "operationId": "CampaignsController_joinCampaign",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/leave/{campaignId}": {
      "delete": {
        "operationId": "CampaignsController_leaveCampaign",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/impression/{token}": {
      "post": {
        "operationId": "CampaignsController_recordImpression",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/click/{token}": {
      "post": {
        "operationId": "CampaignsController_recordClick",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/activate": {
      "post": {
        "operationId": "CampaignsController_activateCampaign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign activated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden - Not campaign owner"
          },
          "404": {
            "description": "Campaign not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate a draft campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/pause": {
      "post": {
        "operationId": "CampaignsController_pauseCampaign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign paused successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden - Not campaign owner"
          },
          "404": {
            "description": "Campaign not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pause an active campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/resume": {
      "post": {
        "operationId": "CampaignsController_resumeCampaign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign resumed successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden - Not campaign owner"
          },
          "404": {
            "description": "Campaign not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resume a paused campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaignId}/leave": {
      "post": {
        "operationId": "CampaignsController_leaveCampaignEarly",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Left campaign successfully, earnings released"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Active participation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Streamer leaves campaign early",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaignId}/pause-participation": {
      "post": {
        "operationId": "CampaignsController_pauseParticipation",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Participation paused successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Active participation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Streamer pauses campaign participation",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaignId}/resume-participation": {
      "post": {
        "operationId": "CampaignsController_resumeParticipation",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "description": "Campaign ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Participation resumed successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Paused participation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Streamer resumes campaign participation",
        "tags": [
          "campaigns"
        ]
      }
    },
...
    "/api/v1/campaigns/completion-check/trigger": {
      "post": {
        "description": "Trigger an immediate check of all campaigns for completion criteria (Admin only)",
        "operationId": "CampaignsController_triggerCampaignCompletionCheck",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Campaign completion check triggered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manually trigger campaign completion check",
        "tags": [
          "campaigns"
        ]
      }
    },
...
  },
  "info": {
    "title": "GameTriggers API",
    "description": "The Gametriggers marketplace API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateCampaignDto": {
        "type": "object",
        "properties": {}
      },
      "JoinCampaignDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}